Ajout des Services

+++ src/main/java/com/Pfe/ecommerce/Back_end/Models/Commentaire.java
@@ -17,7 +17,7 @@ import java.util.Date;
 public class Commentaire {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int idCmt;
+    long idCmt;
     String Commentaire;
     Date datePub;
