+++ src/main/java/com/Pfe/ecommerce/Back_end/Models/Product.java
@@ -18,7 +18,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int idPrd;
+    long idPrd;
 
     String libelle,marque,desc;
     Double prix,fraisex;
