+++ src/main/java/com/Pfe/ecommerce/Back_end/Models/WishList.java
@@ -17,7 +17,7 @@ public class WishList {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int idWL;
+    long idWL;
 
     String libelle;
